Refactor left sidebar and its contents
