+++ pages/market/[[...market]].tsx
@@ -23,6 +23,7 @@ import MobileTrade from 'sections/futures/MobileTrade/MobileTrade';
 import { currentMarketState } from 'store/futures';
 import { RefetchProvider } from 'contexts/RefetchContext';
 import AppLayout from 'sections/shared/Layout/AppLayout';
+import LeftSidebar from '../../sections/futures/LeftSidebar/LeftSidebar';
 
 type AppLayoutProps = {
 	children: React.ReactNode;
@@ -51,7 +52,7 @@ const Market: MarketComponent = () => {
 				<PageContent>
 					<StyledFullHeightContainer>
 						<StyledLeftSideContent>
-							<TradingHistory />
+							<LeftSidebar />
 						</StyledLeftSideContent>
 						<StyledMainContent>
 							<MarketInfo />
