+++ sections/futures/FuturesMarketTabs/constants.ts
@@ -1,0 +1,10 @@
+export const DEFAULT_DATA = [
+	{
+		market: '-',
+		price: '-',
+		volume: '-',
+		fundingRate: '-',
+		priceChange: '-',
+		openInterest: '-',
+	},
+];
