+++ sections/futures/TradingHistory/index.ts
@@ -1,1 +1,1 @@
-export { default } from './TradingHistory';
+export { default } from './TradesHistoryTable';
